Move Outcome class to a separate file
